fix: solve login error and authentication

+++ src/database/migrations/20230908202818_create_users_table.js
@@ -10,7 +10,7 @@ exports.up = function(knex) {
     table.string("email").notNullable();
     table.string("password").notNullable();
     table.string("phoneNumber").notNullable();
-    table.integer("accountNumber").unsigned().notNullable();
+    table.bigInteger("accountNumber").unsigned().notNullable();
     table.timestamps(true, true);
   });
 };
